Refactor binding analysis
