+++ test/unit/main-equal/output.js
@@ -36,12 +36,12 @@ module.exports =
 /******/ 	runtime(__webpack_require__);
 /******/
 /******/ 	// Load entry module and return exports
-/******/ 	return __webpack_require__(__webpack_require__.s = 336);
+/******/ 	return __webpack_require__(__webpack_require__.s = 599);
 /******/ })
 /************************************************************************/
 /******/ ({
 
-/***/ 336:
+/***/ 599:
 /***/ (function(module, __unusedexports, __webpack_require__) {
 
 /* module decorator */ module = __webpack_require__.nmd(module);
